feat: use websockt url path to specify game id
